Fixed Remove Books, now removes one copy

+++ lib_project.py
@@ -334,13 +334,28 @@ def add_book(db, user_id):
 
 
 @get('/librarian/<user_id>/remove/<book_id>')
-def remove_books(db, user_id, book_id):
-    print(book_id)
-    print(user_id)
-    print('removing books')
-    db.execute("""DELETE FROM copy WHERE book_id = ?;""", (book_id,))
-    db.execute("""DELETE FROM book WHERE id = ?;""", (book_id,))
-    redirect(f'/librarian/{user_id}/home')
+def remove_copy(db, user_id, book_id):
+    # Find one copy ID to be removed
+    result = db.execute("""SELECT id FROM copy where book_id = ?;""", (book_id,)).fetchone()
+    copy_id = result['id']
+
+    # Remove associated child loans
+    db.execute("""DELETE FROM loan WHERE copy_id = ?;""", (copy_id,))
+
+    # Remove book if last copy
+    num_copies = db.execute("""SELECT COUNT (copy.id)
+                               FROM copy
+                               WHERE book_id = ?;
+                               """, (book_id,)).fetchone()[0]
+
+    db.execute("""DELETE FROM copy WHERE id = ?;""", (copy_id,))
+    if num_copies == 1:
+        db.execute("""DELETE FROM book WHERE id = ?;""", (book_id,))
+        redirect(f'/librarian/{user_id}/home')
+        return
+
+    redirect(f'/librarian/{user_id}/book/{book_id}')
+
 
 
 @get('/librarian/<user_id>/edit/<book_id>')
